Start error handling

+++ src/routes/message.js
@@ -1,5 +1,7 @@
 import { v4 as uuidv4 } from 'uuid';
 import { request, Router } from 'express';
+import { BadRequestError, NotfoundError } from '../utils/errors';
+
 
 const router = Router();
 
@@ -8,28 +10,39 @@ router.get('/', async (req, res) => {
     return res.send(messages);
 });
 
-router.get('/:messageId', async (req, res) => {
+router.get('/:messageId', async (req, res, next) => {
     const message = await req.context.models.Message.findByPk(
         req.params.messageId,
-    );
+    ).catch((error) => next(new BadRequestError(error)));
+
     return res.send(message);
 });
 
-router.post('/', async (req, res) => {
-  const message = await req.context.models.Message.create({
-    text: req.body.text,
-    userId: req.context.me.id,
-  });
+router.post('/', async (req, res, next) => {
+    const message = await req.context.models.Message.create({
+        text: req.body.text,
+        userId: req.context.me.id,
+    }).catch((error) => next(new BadRequestError(error)));
 
   return res.send(message);
 });
 
 
-router.delete('/:messageId', async (req, res) => {
-  const result = await req.context.models.Message.destroy({
-    where: { id: req.params.messageId },
-  });
-  return res.send(true);
+router.delete('/:messageId', async (req, res, next) => {
+  const element = await req.context.models.Message.findByPk(
+    req.params.messageId,
+  );
+  if (!element) {
+    const newerror = new Error('Not Found');
+    newerror.statusCode = 404;
+    next(newerror);
+  } else {
+    const result = await req.context.models.Message.destroy({
+        where: { id: req.params.messageId },
+    }).catch((error) => next(new BadRequestError(error)));
+    
+    return res.send(true);
+  }
 });
 
 export default router;
